- unexpectedly (as least for me) the HltDecReports in the stripping change 'shape',    i.e. the number of lines which are processed varies (occasionally) from event    to event. Enhance the 'hinted' insertion to deal with this...
git-svn-id: http://svn.cern.ch/guest/lhcb/Phys/trunk@182052 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 82556b4e62fa7a3fdd4226f6d37cbddae717103d

+++ Phys/Kernel/SelectionLine/Kernel/SelectionLine.h
@@ -172,7 +172,7 @@ namespace Selection {
       int  m_maxAcceptOnError;               ///< quota to avoid runaway accepts in case of persistent errors..
       int  m_nAcceptOnError;
       unsigned m_slowThreshold;
-      int m_insertion_hint = -1;
+      size_t m_insertion_hint = -1;
     };
 }
 #endif // Line_H
